feat($search): Search in all pages with posts
- implements PostsMain (todo: change this name), as a component.
- the PostsMain has categories, tags and search, surrounded with props
- component used in Home, Posts, Category and Tag layouts
